refactor(sdk): unify nowTimestamp implementation
Extract nowTimestamp into sdk/now-timestamp.cjs as single source of truth,
used by both runner-io (ESM) and ffi_js.mbt (CJS require).

Resolves tornado-80h
