+++ sdk/runner-io.mjs
@@ -1,10 +1,6 @@
-function nowTimestamp() {
-  const d = new Date();
-  const h = String(d.getHours()).padStart(2, "0");
-  const m = String(d.getMinutes()).padStart(2, "0");
-  const s = String(d.getSeconds()).padStart(2, "0");
-  return `${h}:${m}:${s}`;
-}
+import { createRequire } from "node:module";
+const require = createRequire(import.meta.url);
+export const { nowTimestamp } = require("./now-timestamp.cjs");
 export function stampEvent(event) {
   if (event != null && typeof event === "object") {
     return { ...event, _tornado_ts: nowTimestamp() };
